style(cycle-status): refactors components to move calls upward
